Allow chaining commands after VtrAttachToPane
Occasionally it can be useful to be able to interact with more than one
REPL at a time. VTR doesn't have official support for this behavior, but
we can combine the `VtrAttachToPane` command and the
`VtrSendLinesToRunner` command to provide rough support for this.

This change updates the `VtrAttachToPane` configuration to support this
chaining.

*Note* this mechanism will only allow for sending single lines.

``` vim
nmap <leader>v1 :VtrAttachToPane 1 \| VtrSendLinesToRunner<cr>
nmap <leader>v2 :VtrAttachToPane 2 \| VtrSendLinesToRunner<cr>
```